Extract some interpreter related things out into a separate module
