+++ gibbon-compiler/src/Gibbon/L1/Interp.hs
@@ -32,6 +32,7 @@ import qualified Data.ByteString.Lazy.Char8 as B
 import qualified Data.Sequence as S
 
 import           Gibbon.Common
+import           Gibbon.Interp
 import           Gibbon.GenericOps
 import           Gibbon.L1.Syntax as L1
